Created a small lib for a dual purpose control knob
This first version uses a potentiometer to track a value. If the value
rises above a certain threshold, the library indicates a mode change. It
keeps information whether the knob is currently used to control mode 1 or
mode 2, making it possible to control two analog values using only one
knob.
At the moment, the library is not time sensitve, which means that in order
to change the mode, one of the values is constantly altered until the
threshold is reached. This is of course undesirable and will be addressed
in a future version.