Prepared sketch for RTC and data logging
The pins used needed to be moved in order to use the Adafruit datalogging
shield. In addition, a DFC77 receiver (European atom clock) will be used
for precise timestamps.

I removed the panel voltage for the time being because of the common
ground issue. Maybe I'll add it again later.