Added second battery voltage, disabled DCF
I added a second battery voltage to be measured and had to disable the DCF
part of the code due to a conflict between the DCF stuff and the SD card.
It somehow got impossible to write to a file and only removing the DCF
code resolved that. I will have to investigate this further.

Be warned: work in progress